requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
blinker==1.9.0
cachetools==6.1.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
contourpy==1.3.2
cycler==0.12.1
exceptiongroup==1.3.0
fastapi==0.116.1
filelock==3.19.1
fonttools==4.59.1
fsspec==2025.7.0
gitdb==4.0.12
GitPython==3.1.45
h11==0.16.0
idna==3.10
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
kiwisolver==1.4.9
MarkupSafe==3.0.2
matplotlib==3.10.5
mpmath==1.3.0
narwhals==2.1.2
networkx==3.4.2
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.1
pillow==11.3.0
protobuf==6.32.0
psutil==7.0.0
py-cpuinfo==9.0.0
pyarrow==21.0.0
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.0
scipy==1.15.3
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.47.2
streamlit==1.48.1
streamlit-lottie==0.0.5
sympy==1.14.0
tenacity==9.1.2
toml==0.10.2
torch==2.8.0
torchvision==0.23.0
tornado==6.5.2
tqdm==4.67.1
triton==3.4.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
ultralytics==8.3.181
ultralytics-thop==2.0.15
urllib3==2.5.0
uvicorn==0.35.0
watchdog==6.0.0
